move purger to different directory
